implement UserService class
