+++ src/main/java/com/shelterapp/backend/users/UserController.java
@@ -1,0 +1,35 @@
+package com.shelterapp.backend.users;
+
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/users")
+public class UserController {
+
+    private final UserService userService;
+
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @GetMapping
+    public List<User> getUsers() {
+        return userService.findAll();
+    }
+
+    @GetMapping("/{id}")
+    public User getUser(@PathVariable Long id) {
+        return (User) userService.getUser(id);
+    }
+
+	@PostMapping
+	public void registerNewUser(@RequestBody User user) {
+		userService.registerNewUser(user);
+	}
+
+}
